% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialise_state.R
\name{initialise_state}
\alias{initialise_state}
\title{Initialise evaluation state}
\usage{
initialise_state(data, context, criteria, options, extras = list())
}
\arguments{
\item{data}{data.frame/tibble of APSIM outputs.}

\item{context}{list controlling variable mapping and reporting context.}

\item{criteria}{list controlling decision criteria.}

\item{options}{list controlling output toggles and figure behaviour.}

\item{extras}{additional metadata stored in the state object.}
}
\description{
Initialise evaluation state
}
